Fix concurrent mutation of attributes in options
WithInstrumentationAttributes is creating a closure with
a reference to a slice which is later passed to attribute.NewSet.

attribute.NewSet may mutate the slice, so this will lead to
a data race when the option is applied concurrently. We can
fix this by moving the call to attribute.NewSet outside the
closure.