Generate gRPC Client target parsing func
In order to add gRPC server attributes for exporter observability, this
information needs to be parsed into a host and port. The added generated
files provides ParseCanonicalTarget for this functionality.

This is added as a generated template as it is expected to be needed for
all OTLP exporters (e.g. #7404, #7353).