Accept gRPC status code as param to End
The client is already parsing the error for the status code. Do not
duplicate the work, instead accept that parsed value as an argument when
producing telemetry.

The status.Code function is heavy. It requires 3 allocations of
significant memory and incurs a fair amount of computational overhead.
Reducing the number of times it is called is a performance win.